Tests for node conversion
Moved Node conversion tests into `node_conversion_test.go` to align
better with what is being tested.

Added a couple of tests to validate annotations are being added/removed
when merging a Calico Node into a K8s Node.
